Added a method for counting roll to escape jail attempts.
Refactored methods in the MonopolyEngine.

Refactored when a player has to pay to get out of jail.

A player will remain in jail for three turns if they do not roll doubles, test passes.

If a player fails to roll doubles for three turns they must pay 50, test passes.

If a player is in jail and they roll doubles then they move that many spaces, test passes.

If a player is in jail and they roll doubles then they do not roll again.